requirements.txt
Pillow
Pygments
Markdown==2.6.8
Pygments==2.2.0
pytz==2016.10
setuptools-scm
django-haystack==2.6.1
jieba==0.38
Whoosh==2.7.4